[#7] Uncomment code;
